requirements.txt
scipy==1.7.0
matplotlib==3.3.4
numpy==1.20.1
statsmodels==0.12.2
torch==1.9.0
pandas==1.2.4
python_igraph==0.9.6
scanpy==1.8.1
